@@ -3,15 +3,30 @@
     "license": "MIT",
     "type": "novosga-module",
     "autoload": {
-        "psr-4": { "Novosga\\TriageBundle\\": "" }
+        "psr-4": { "Novosga\\TriageBundle\\": "src/" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Novosga\\TriageBundle\\Tests\\": "tests/" }
     },
     "require": {
-        "php": ">=7.2.5",
-        "novosga/core": "^2.1"
+        "php": ">=8.2",
+        "doctrine/orm": "^3.2",
+        "novosga/core": "self.version",
+        "symfony/framework-bundle": "7.1.*"
+    },
+    "require-dev": {
+        "phpstan/phpstan": "^1.10",
+        "phpunit/phpunit": "^9.5",
+        "slevomat/coding-standard": "~8.0"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.1-dev"
+            "dev-master": "2.2-dev"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
         }
     }
 }
